requirements.txt
flask
pandas
scikit-learn
google-generativeai
gunicorn
python-dotenv
werkzeug
jinja2
itsdangerous
click